Backend/python/requirements.txt
openai
mysql-connector-python
langchain==0.0.292
pandas==2.1.0
cryptoconditions
google-cloud-secret-manager==2.0.0
